Isolate out the filesystem parts of the url resolver from the package relative parts.
